+++ front/app/containers/Admin/settings/messages.ts
@@ -123,10 +123,10 @@ export default defineMessages({
     id: 'app.containers.AdminPage.SettingsPage.contentModeration',
     defaultMessage: 'Content moderation',
   },
-  registrationTabDescription: {
-    id: 'app.containers.AdminPage.SettingsPage.registrationTabDescription',
+  registrationHelperTextDescription: {
+    id: 'app.containers.AdminPage.SettingsPage.registrationHelperTextDescription',
     defaultMessage:
-      'Provide a short description at the top of your registration form. Add and configure the registration fields for your platform.',
+      'Provide a short description at the top of your registration form.',
   },
   profanityBlockerSetting: {
     id: 'app.containers.AdminPage.SettingsPage.profanityBlockerSetting',