+++ front/app/modules/commercial/user_custom_fields/admin/components/CustomFields/All/messages.ts
@@ -27,4 +27,13 @@ export default defineMessages({
     id: 'app.containers.AdminPage.SettingsPage.CustomSignupFields.customFieldDeletionConfirmation',
     defaultMessage: 'Are you sure you want to delete this custom field?',
   },
+  customFieldsTooltip: {
+    id: 'app.containers.AdminPage.SettingsPage.CustomSignupFields.customFieldsTooltip',
+    defaultMessage:
+      'Drag and drop the fields to determine the order in which they appear in the sign-up form.',
+  },
+  customFieldsSubSectionTitle: {
+    id: 'app.containers.AdminPage.SettingsPage.CustomSignupFields.customFieldsSubSectionTitle',
+    defaultMessage: 'Fields',
+  },
 });